--- server/tests/test.test.js
@@ -1,11 +1,0 @@
-/* eslint-disable linebreak-style */
-import { assert } from 'chai';
-
-
-describe('A feature', () => {
-    describe('A scenario', () => {
-        it('do something expexted', () => {
-            assert(true);
-        });
-    });
-});
